back/requirements.txt
Django>=4.2
python-decouple
boto3==1.34.46
botocore==1.34.46
PyJWT==2.7.0
bcrypt==3.2.2
cryptography==41.0.7
requests==2.31.0
gunicorn==21.2.0
pytz==2024.1
django-cors-headers==4.3.1
mysqlclient==2.2.1
PyYAML==6.0.1
social-auth-app-django>=5.2.0
djangorestframework-simplejwt==5.3.1
google-auth>=2.0.0,<3.0.0
Pillow>=10.0.0